feat: configfileparsing cleaner version
