+++ config/wrong/configfile_misplaced_semicolon
