+++ config/wrong/configfile_missing_semicolon
