+++ config/wrong/configfile_to_many_closing_brackets
