+++ config/wrong/configfile_unclosed-bracket
