+++ config/wrong/configfile_wrong_brackets
